project executable files/requirements.txt
streamlit
pandas
langchain
langchain-google-genai
google-generativeai
openpyxl
